feat(auth): add refreshing tokens flow (#249)
### What does this PR do?
This PR adds a feature to refresh tokens, so a user is not logged out
after their token expires.

### Changes brought by this PR
- The main change is `AuthInterceptor `, it catches 401 errors from BE,
blocks other requests (because it's a `QueuedInterceptor`) and tries to
refresh token. After that it sets a new token header.

- Added `DelegatedLoginService` with its dependencies to
ion_identity_client. This thing makes the request to get a new token.
- Updated ion_identity_client/example a little bit.
- Added `AppLifecycleProvider`, it holds the `AppLifecycleState`