fix: networks in send/receive coins from profile modals (#791)
## Description
1. Fixed sorting issue (Networks were ignored during comparison).
2. Attached coins, networks data to the implemented UI for the
receive/send coins from in the profile tab.

P.S. As for now, these pages almost the same, so you can see some sort
of copy/paste there. I think, it will be changed when we will actually
implement these flows for the profile. As for now, the task was only to
use real wallets-related data.

## Type of Change
- [x] Bug fix
- [x] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore

## Screenshots (if applicable)

https://github.com/user-attachments/assets/c87be7f9-40b9-44a7-a4a1-716b4ddd7b28

---------

Co-authored-by: ice-tychon <175587874+ice-tychon@users.noreply.github.com>