feature(wallet): show NFTs on Wallet page (#824)
## Description
This PR adds displaying NFTs on Wallet page. It has a new proposal for
state management, MVVM.
- `NftRepository` receives NFTs for a specified
`walletViewId`/`walletId`. In future it will be able to cache data.
- `LoadNftsUseCase` loads all NFTs for a specified `walletViewId` using
`NftsRepository`.
- `WalletNftsViewModel` contains everything the Wallet page needs to
display NFTs. It uses "commands" to receive events from Widgets and hold
result of an action, and listenables to store state which is not a
direct result of a command.
- Widgets use `CommandBuilder` or `ValueListenableBuilder` to draw UI
for a state in a ViewModel.
- Other changes are just to make the project compile. Those pages are
gonna be implemented in later PRs.

## Type of Change
- [ ] Bug fix
- [x] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore

## Screenshots
![photo_2025-03-07_14-46-12
(2)](https://github.com/user-attachments/assets/702d8065-5f29-449a-8f2f-4e09421e4dba)

![photo_2025-03-07_14-46-12](https://github.com/user-attachments/assets/4ca7b4ce-f7e1-459b-8615-52974c9099db)