feat: pause system playback on video unmute only (#984)
## Description
Pause system playback only when a video in the app is unmuted. And allow
it back what a vdeo in the app is muted again.
It is achieved by setting mixWithOthers to isMuted value. But that can
be set only during video controller creation. So I changed
videoControllerProvider to return
Future<Raw<CachedVideoPlayerPlusController>> in order to return only
fully baked and initialized controller and in case of recreation when
globalMuteProvider is changed - keep the position and playback status of
current one.

## Type of Change
- [ ] Bug fix
- [x] New feature
- [ ] Breaking change
- [x] Refactoring
- [ ] Documentation
- [ ] Chore