fix: stop video from pausing in full screen when mute is toggled (#1071)
## Description
This PR fixes an issue where the video in full screen mode would pause
when muted/unmuted

## Additional Notes
This issue is actually caused by the video_preview present inside feed
cell underneath that uses the same video controller. It has a `onEffect`
which pauses the video, because cell is not visible on the screen. By
introducing a `wasOffScreen` flag we avoid video pausing after
controller changes if the video_preview was already off screen before
that.

## Type of Change
- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore

## Screenshots (if applicable)
<!-- Include screenshots to demonstrate any UI changes. -->
<!-- <img width="180" alt="image" src="image_url_here"> -->