fix(wallet): show FriendsList only if at least 1 friend's metadata is loaded (#1095)
## Description
Previously, the Friends section on the Wallet page was shown right after
the friends' pubkeys are loaded. This lead to a situation when the
section is shown, but friends are just Loaders.

This PR changes the way we show the Friends list. We show a Friends
section loader until at least 1 of the friend's metadata is loaded.

## Type of Change
- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore

## Screenshots
![CleanShot 2025-04-26 at 19 16
18](https://github.com/user-attachments/assets/1ed1665f-8ae2-4982-95fe-aaea08616ed3)