refactor: extract FeedGlobalNotifications from GlobalNotifications (#1139)
## Description
This PR makes Global Notifications more global. Previously they were
tightly coupled to the feed.

### Changes
- Extracted `GlobalNotificationView` from `GlobalNotificationBar`
- `createPostNotifierProvider` listeners extracted from
`GlobalNotificationBar` to a separate file in the Feed feature.
- Changed `GlobalNotificationNotifier` state type and fixed possible
race condition with a `CancelableOperation`
- Introduced `GlobalNotification` type along with
`FeedGlobalNotification` instead of the `GlobalNotificationData` to make
it possible to reuse global notifications for other purposes, not only
for Feed.
- Added `equatable` dependency (it already was in the project, but now
it's direct, not transitive).

## Type of Change
- [ ] Bug fix
- [ ] New feature
- [ ] Breaking change
- [x] Refactoring
- [ ] Documentation
- [ ] Chore

## Screenshots


https://github.com/user-attachments/assets/4c40de09-3753-48ef-8073-2415a1bbf8b8