feat: passkeys autosuggestions (#1182)
## Description
This PR adds the autosuggestion feature for Passkeys on Login. It
suggests passkeys only once per login session (if a user cancels the
dialog, the app will show it again only on the next start or after
logout).

### Changes
- Renamed `GetStartStep` to `SignInStep`
- Refactored `SignInStep` to include the `onLogin` code (moved the code
from `GetStartedPage`). Now the `SignInStep` contains all the required
code for both regular sign in and sign in with suggestion.
- Added required code to extract username from token, because in case we
login with autosuggestion, we don't know the username to save a token
for it.

## Type of Change
- [ ] Bug fix
- [x] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore

## Screenshots

https://github.com/user-attachments/assets/985ab02e-7e10-4c84-beaa-0c32c77fee69