fix: make coin amount input field the same as on the web (#1186)
## Description
So, the main purpose of new text formatter is to use comma as delimiter
between int digits. So, if I input `9999`, then the text field should
show `9,999`. So complex logic is necessary, cos iOS, for example, use
comma as delimiter between ints and decimals, so we need to handle such
cases.

Fixed bugs:
- Removed rounding for the coin icons.
- Used correct formatting for the coin balance on the `CoinDetails`
page.

## Type of Change
- [x] Bug fix
- [x] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore

## Screenshots (if applicable)

https://github.com/user-attachments/assets/94ff3d1a-fc28-4993-963a-ed73150bbdbe