fix: support transfers from one wallet view to another (#1202)
## Description
Bind transactions to the wallet view id in the DB.
Please, note, that we changed the primary key of the transactions table
in the db, so we need to recreate it. I've added a simple migration
logic in the repository to add already stored data to the transactions
in the table.

## Type of Change
- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore