feat: set SizedBox as an error widget for release builds (#1205)
## Description
This PR sets a SizedBox as an ErrorWidget that will be rendered during
build method exceptions in release mode.

## Additional Notes
This change prevents parts of the app from not being usable in case of
any build error. For example, when exception was thrown during build
method of a post cell on feed, whole feed became unusable. This change
is going to prevent that from happening.

## Type of Change
- [ ] Bug fix
- [x] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore

## Screenshots (if applicable)
Sample screenshot showcasing how feed will look like when post body
fails to render.
![ScreenShot 2025-05-14 at 11 00
03@2x](https://github.com/user-attachments/assets/808508bf-2d23-49f7-8125-5ab9a946dd06)