feat: global long running subscription (#1469)
## Description
This PR adds a single global long-running subscription and removes
long-running subscription creation from other parts of the codebase. To
achieve this,
-  Subscribe to event messages that require immediate updates. 
- Save the latest fetched time in local storage to ensure accurate
fetching of new items on each launch.
- Create two separate filters: one for regular events and another for
encrypted events. For encrypted events, the app provides an overlap
since their createdAt is randomly generated within a 2-day period.
- Add sub-handlers for each flow, which detect whether an event is
related to its own flow and, if so, run the callback.

## Additional Notes
- Refactor 30003 so the app codebase has only one entity class for it
- Link archived_conversations bookmark to 10003.
- Manage bookmarks flow using traditional request/response pattern

## Task ID
ION-2932

## Type of Change
- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [x] Refactoring
- [ ] Documentation
- [ ] Chore

## Screenshots (if applicable)
<!-- Include screenshots to demonstrate any UI changes. -->
<!-- <img width="180" alt="image" src="image_url_here"> -->