feat: set ion connect inside event signer name key (#1482)
## Description
This PR changes naming of secure storage key for event signer to
ion_connect. Also it changes some imports to use ion connect instead of
nostr.

## Additional Notes
This change will cause app to recreate the event signer as it will not
have access to the old one. This will result in a `link new device`
modal to be displayed for all users.

## Task ID
<!-- Add corresponding task ID here. -->

## Type of Change
- [ ] Bug fix
- [x] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore

## Screenshots (if applicable)
<!-- Include screenshots to demonstrate any UI changes. -->
<!-- <img width="180" alt="image" src="image_url_here"> -->