fix(chat): prevent displaying blurhash when local file is available in visual media message component (#1495)
## Description
This PR fix stops blurhash from displaying when a local file is
available in the visual media message component. As a result, in the
stack, if we use blurhash under the image and it has a transparent area,
it appears as if it hasn't loaded.

## Additional Notes
N/A

## Task ID
ION-3006

## Type of Change
- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore

## Screenshots (if applicable)
<!-- Include screenshots to demonstrate any UI changes. -->
<img width="180" alt="image"
src="https://github.com/user-attachments/assets/4820df5a-8dba-41af-87ce-6d667f42b803">