fix: likes sync issue (#1512)
## Description
Fixed a corner case in a post likes counter sync.
Here is the flow:
- Like a post with 0 likes and you will see results through optimistic
UI;
- Make pull to refresh for that post and new data is fetched including
likes counter entity - so you see the correct result now from real data;
- Press like button again to remove you like from the post - you will
see 0 likes immediately thanks to optimistic UI;
- Make pull to refresh for that post and new data is fetched including
likes counter entity - now you will see post has 1 like. That is because
optimistic UI is reset on pull to refresh but cached entity with
previous likes counter is not replaced by new one because the post now
has no likes so no entity is returned on pull to refresh and we reuse
the old cached one on UI and display 1 like for the post;

So added cache cleaning for the post likes entity on like toggle;

## Type of Change
- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore

## Screenshots (if applicable)


https://github.com/user-attachments/assets/8f660a2b-3bbc-440f-bd3b-410d3782a061