fix: don't hide MoneyMessage when FundsRequest is cancelled (#1623)
## Description
Currently when a user pressed "Cancel" button on a MoneyMessage, the
message disappears. This PR changes this behavior so the funds request
message stays in the conversations but "cancel" and "send" buttons are
disabled.

To do this I send kind 5 for the 1755 event and update it in the db.

## Task ID
ION-3082

## Type of Change
- [x] Bug fix

## Screenshots


https://github.com/user-attachments/assets/eb62f351-4942-45a3-bec2-dc0af2d1df20