feat: introduce explore feed modifier type (#1619)
## Description
This PR introduces `ExploreModifierType` with 3 values:
* `withAnyTopic` - Fetch events with any topic specified. Default
behavior.
* `interests` - Fetch events with the provided tag. Is used for
Articles, if we have at least 1 selected article category.
* `any` - Fetch all events, with or without tags. Is used if
`excludeUnclassifiedFromExplore` in feed remote config is `true`.

## Additional Notes
* Fixes seen events querying from DB - query only events from users that
we're currently following. Fixes the case when we follow some user, get
some events to the seen cache, unfollow this user, we still see the seen
events from that user on feed.

## Task ID
ION-3194

## Type of Change
- [x] Bug fix
- [x] New feature
- [ ] Breaking change
- [x] Refactoring
- [ ] Documentation
- [ ] Chore