feat: read write relays (#1627)
## Description
This PR introduces `read` and `write` relay modifiers. To send events,
we pick a random `write` relays from the user list. For the "read" flow,
everything stays the same.

## Additional Notes
* Handle `relay-is-read-only` relay exception. Is returned on attempt to
send an event to `read-only` relay. In this case we mark the used relay
locally as `read-only` and retrying with other ones. If no `write`
relays left, refetch the list.
* Rename relay_creation_provider -> relay_picker_provider, since the
main responsibility is to pick the best relay for the provider
`ActionSource` and `ActionType`
* Rework sync chat/user/files-storage relays.


## Task ID
ION-3264

## Type of Change
- [ ] Bug fix
- [x] New feature
- [ ] Breaking change
- [x] Refactoring
- [ ] Documentation
- [ ] Chore