fix(wallet): jumping Wallet not found dialog (#1698)
## Description
This PR fixes an issue where the “Wallet not found” screen would appear
with a jumping effect.
The reason was that the user was initially redirected to the “Share”
dialog, which would then determine there was nothing to share and send
the user to “Wallet not found.” Now, the check for an existing wallet
happens before opening the “Share” dialog, preventing this unnecessary
navigation.

The video in the PR demonstrates the fixed behavior. For details about
the original bug, please refer to the task ticket.

## Task ID
ION-3356

## Type of Change
- [x] Bug fix

## Screenshots

https://github.com/user-attachments/assets/86f0b85c-9fbb-4a49-baa2-0bba165b0906