fix: post sharing as a story increments repost counter (#1713)
## Description
When users share a post as a story by tapping "Add to story", the repost
counter increments. This happens because the story creation flow passes
the quotedEvent parameter, which creates a quote relationship and
increases the quotesCount. Since the repost counter displays
repostsCount + quotesCount, this causes the unintended increment.

Changes:

- Modified StoryPreviewPage to pass sourcePostReference instead of
quotedEvent for post screenshots
- Added sourcePostReference field to ModifiablePostData to store
reference to original post
- Implemented "a" tag with "mention" marker to track source post without
affecting counters
- Updated EntityDataWithRelatedEvents.fromTags() to properly handle "a"
tags with "mention" marker
- Removed ReferencesSearchExtension filter from buildStoriesDataSource
to allow stories with sourcePostReference to be displayed

## Task ID
<!-- Add corresponding task ID here. -->
3357

## Type of Change
- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [x] Refactoring
- [ ] Documentation
- [ ] Chore

## Screenshots (if applicable)
<!-- Include screenshots to demonstrate any UI changes. -->
<!-- <img width="180" alt="image" src="image_url_here"> -->