refactor: stream based relay ranking (#1732)
## Description
This PR changes the way relay ranking returns results. Previously we
were waiting for all the relays to be pinged to return the results. That
was hanging the app if some relay is not responding. Now `.rank` returns
a `Stream` of results, so the app could proceed as soon as the fastest
relay is pinged.

## Additional Notes
* Handles the relay retry for `RelayUnreachableException` explicitly

## Task ID
ION-3369

## Type of Change
- [x] Bug fix
- [x] New feature
- [ ] Breaking change
- [x] Refactoring
- [ ] Documentation
- [ ] Chore