fix: remove image precache from the CoinsGroupItem (#1752)
## Description
`CoinsGroupItem` is used on different pages, like `Receive coins`, `Send
coins` modals, etc. These pages have a search. For example, a user would
paste the word "ice" in the search and see 100 different coins with the
word "ice" in the name, symbol, or something else.

Because of the line
`ref.read(networkSelectorNotifierProvider(symbolGroup:
coinsGroup.symbolGroup));`
for each of these 100 coins
`ref.watch(syncedCoinsBySymbolGroupProvider(symbolGroup));`
would be triggered, which would result in the mobile client making
hundreds (or thousands) of requests in one session.

## Task ID
ION-3460

## Type of Change
- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore